Rename utils to source_location
