+++ include/dlaf/common/assert.h
@@ -16,7 +16,7 @@
 #include <iostream>
 #include <sstream>
 
-#include "dlaf/common/utils.h"
+#include "dlaf/common/source_location.h"
 
 namespace dlaf {
 namespace internal {
