+++ include/dlaf/communication/error.h
@@ -18,7 +18,7 @@
 
 #include <mpi.h>
 
-#include "dlaf/common/utils.h"
+#include "dlaf/common/source_location.h"
 
 namespace dlaf {
 namespace internal {
